video-05&06 with adujest viewModel and partial views

+++ C44-G01-MVC04.BLL/Services/EmployeesServices/EmployeeServices.cs
@@ -1,4 +1,5 @@
 ﻿using AutoMapper;
+using C44_G01_MVC04.BLL.Common.Services.Attachments;
 using C44_G01_MVC04.BLL.Dto_s;
 using C44_G01_MVC04.BLL.Dto_s.EmployeeDto_s;
 using C44_G01_MVC04.BLL.Factories.DepartmentFactory;
@@ -18,11 +19,13 @@ namespace C44_G01_MVC04.BLL.Services.EmployeesServices
     public class EmployeeServices : IEmployeeServices
     {
         private readonly IUnitOfWork unitOfWork;
+        private readonly IAttachmentServices attachmentServices;
         private readonly IMapper mapper;
 
-        public EmployeeServices(IUnitOfWork unitOfWork,IMapper mapper)
+        public EmployeeServices(IUnitOfWork unitOfWork, IAttachmentServices attachmentServices,IMapper mapper)
         {
             this.unitOfWork = unitOfWork;
+            this.attachmentServices = attachmentServices;
             this.mapper = mapper;
         }
         public IEnumerable<EmployeeDto> GetAllEmployee()
@@ -70,6 +73,12 @@ namespace C44_G01_MVC04.BLL.Services.EmployeesServices
             emp.CreatedOn = DateTime.Now;
             emp.LastModifiedBy = 1;
             emp.LastModifiedOn = DateTime.Now; 
+
+            if (dto.Image != null)
+            {
+                emp.ImageName = attachmentServices.UploadImage(dto.Image, "images");
+            }
+
             unitOfWork.EmployeeRepository.Add(emp);
             return unitOfWork.Complete();
         }
@@ -79,12 +88,32 @@ namespace C44_G01_MVC04.BLL.Services.EmployeesServices
             var emp = mapper.Map<UpdatedEmployeeDto, Employee>(dto);
             emp.LastModifiedBy = 1;
             emp.LastModifiedOn = DateTime.Now;
+           
+            if (dto.Image != null)
+            {
+                if (dto.ImageName is not null)
+                {
+                    var oldImagePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "files", "images", emp.ImageName);
+                    attachmentServices.DeleteImage(oldImagePath);
+                }
+                emp.ImageName = attachmentServices.UploadImage(dto.Image, "images");
+            }
+
             unitOfWork.EmployeeRepository.Update(emp);
             return unitOfWork.Complete();
         }
         public int DeleteEmployee(int id)
         {
-            unitOfWork.EmployeeRepository.Delete(id);
+            var emp = unitOfWork.EmployeeRepository.GetById(id);
+            if (emp is not null)
+            {
+                if (emp.ImageName is not null)
+                {
+                    var imagePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot", "files", "images", emp.ImageName);
+                    attachmentServices.DeleteImage(imagePath);
+                }
+                unitOfWork.EmployeeRepository.Delete(id);
+            }
             return unitOfWork.Complete();
         }
     }
